requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
pymongo==4.4.0
dnspython==2.4.2
requests==2.31.0
python-dotenv==1.0.1
Werkzeug==2.3.7
gunicorn==21.2.0